move out http2 from curl, remove grpc and hapi and do it in a seperate pr
