--- packages/datadog-plugin-grpc/test/integration-test/client.spec.js
@@ -1,52 +1,0 @@
-'use strict'
-
-const {
-  FakeAgent,
-  createSandbox,
-  checkSpansForServiceName,
-  skipUnsupportedNodeVersions,
-  spawnPluginIntegrationTestProc
-} = require('../../../../integration-tests/helpers')
-const { assert } = require('chai')
-
-const describe = skipUnsupportedNodeVersions()
-
-describe('esm', () => {
-  let agent
-  let proc
-  let sandbox
-
-  before(async function () {
-    this.timeout(20000)
-    sandbox = await createSandbox(['@grpc/grpc-js', '@grpc/proto-loader', 'get-port'], false, [
-      `./packages/datadog-plugin-grpc/test/*`])
-  })
-
-  after(async () => {
-    await sandbox.remove()
-  })
-
-  beforeEach(async () => {
-    agent = await new FakeAgent().start()
-  })
-
-  afterEach(async () => {
-    proc && proc.kill()
-    await agent.stop()
-  })
-
-  context('grpc', () => {
-    it('is instrumented', async () => {
-      const res = agent.assertMessageReceived(({ headers, payload }) => {
-        console.log(headers, payload)
-        assert.propertyVal(headers, 'host', `127.0.0.1:${agent.port}`)
-        assert.isArray(payload)
-        assert.strictEqual(checkSpansForServiceName(payload, 'grpc.client'), true)
-      })
-
-      proc = await spawnPluginIntegrationTestProc(sandbox.folder, 'integration-test/server.mjs', agent.port)
-
-      await res
-    }).timeout(20000)
-  })
-})
