+++ packages/datadog-plugin-http2/test/integration-test/client.spec.js
@@ -3,11 +3,11 @@
 const {
   FakeAgent,
   createSandbox,
-  curlAndAssertMessage,
   skipUnsupportedNodeVersions,
   spawnPluginIntegrationTestProc
 } = require('../../../../integration-tests/helpers')
 const { assert } = require('chai')
+const http2 = require('http2')
 
 const describe = skipUnsupportedNodeVersions()
 
@@ -19,7 +19,7 @@ describe('esm', () => {
   before(async function () {
     this.timeout(20000)
     sandbox = await createSandbox(['http2'], false, [
-      `./packages/datadog-plugin-http2/test/*`])
+      `./packages/datadog-plugin-http2/test/integration-test/*`])
   })
 
   after(async () => {
@@ -37,9 +37,8 @@ describe('esm', () => {
 
   context('http2', () => {
     it('is instrumented', async () => {
-      proc = await spawnPluginIntegrationTestProc(sandbox.folder, 'integration-test/server.mjs', agent.port)
-
-      return curlAndAssertMessage(agent, proc, ({ headers, payload }) => {
+      proc = await spawnPluginIntegrationTestProc(sandbox.folder, 'server.mjs', agent.port)
+      const resultPromise = agent.assertMessageReceived(({ headers, payload }) => {
         assert.propertyVal(headers, 'host', `127.0.0.1:${agent.port}`)
         assert.isArray(payload)
         assert.strictEqual(payload.length, 1)
@@ -47,7 +46,41 @@ describe('esm', () => {
         assert.strictEqual(payload[0].length, 1)
         assert.propertyVal(payload[0][0], 'name', 'web.request')
         assert.propertyVal(payload[0][0].meta, 'component', 'http2')
-      }, undefined, true)
+      })
+      await curl(proc)
+      return resultPromise
     }).timeout(20000)
   })
 })
+
+async function curl (url) {
+  if (typeof url === 'object') {
+    if (url.then) {
+      return curl(await url)
+    }
+    url = url.url
+  }
+
+  const urlObject = new URL(url)
+  return new Promise((resolve, reject) => {
+    const client = http2.connect(urlObject.origin)
+    client.on('error', reject)
+
+    const req = client.request({
+      ':path': urlObject.pathname,
+      ':method': 'GET'
+    })
+    req.on('error', reject)
+
+    const chunks = []
+    req.on('data', (chunk) => {
+      chunks.push(chunk)
+    })
+
+    req.on('end', () => {
+      resolve(Buffer.concat(chunks))
+    })
+
+    req.end()
+  })
+}
